+++ lib/l10n/app_localizations_en.dart
@@ -35850,6 +35850,14 @@ class AppLocalizationsEn extends AppLocalizations {
   }
 
   @override
+  String notificationChannelMessageTitleNoCode(
+    String senderName,
+    String channelName,
+  ) {
+    return '$senderName in $channelName';
+  }
+
+  @override
   String notificationMeshCoreAdvertTitle(String contactName, String shortCode) {
     return 'New peer: $contactName ($shortCode)';
   }