+++ lib/l10n/app_localizations_es.dart
@@ -36431,6 +36431,14 @@ class AppLocalizationsEs extends AppLocalizations {
   }
 
   @override
+  String notificationChannelMessageTitleNoCode(
+    String senderName,
+    String channelName,
+  ) {
+    return '$senderName en $channelName';
+  }
+
+  @override
   String notificationMeshCoreAdvertTitle(String contactName, String shortCode) {
     return 'Nuevo par: $contactName ($shortCode)';
   }