+++ lib/l10n/app_localizations_it.dart
@@ -36266,6 +36266,14 @@ class AppLocalizationsIt extends AppLocalizations {
   }
 
   @override
+  String notificationChannelMessageTitleNoCode(
+    String senderName,
+    String channelName,
+  ) {
+    return '$senderName in $channelName';
+  }
+
+  @override
   String notificationMeshCoreAdvertTitle(String contactName, String shortCode) {
     return 'Nuovo peer: $contactName ($shortCode)';
   }