+++ lib/l10n/app_localizations_pt.dart
@@ -36239,6 +36239,14 @@ class AppLocalizationsPt extends AppLocalizations {
   }
 
   @override
+  String notificationChannelMessageTitleNoCode(
+    String senderName,
+    String channelName,
+  ) {
+    return '$senderName em $channelName';
+  }
+
+  @override
   String notificationMeshCoreAdvertTitle(String contactName, String shortCode) {
     return 'Novo peer: $contactName ($shortCode)';
   }