+++ lib/l10n/app_localizations_ru.dart
@@ -36274,6 +36274,14 @@ class AppLocalizationsRu extends AppLocalizations {
   }
 
   @override
+  String notificationChannelMessageTitleNoCode(
+    String senderName,
+    String channelName,
+  ) {
+    return '$senderName в $channelName';
+  }
+
+  @override
   String notificationMeshCoreAdvertTitle(String contactName, String shortCode) {
     return 'Новый пир: $contactName ($shortCode)';
   }