+++ lib/l10n/app_localizations_uk.dart
@@ -36285,6 +36285,14 @@ class AppLocalizationsUk extends AppLocalizations {
   }
 
   @override
+  String notificationChannelMessageTitleNoCode(
+    String senderName,
+    String channelName,
+  ) {
+    return '$senderName у $channelName';
+  }
+
+  @override
   String notificationMeshCoreAdvertTitle(String contactName, String shortCode) {
     return 'Новий пер: $contactName ($shortCode)';
   }